+++ sunpy/instr/aia.py
@@ -44,11 +44,11 @@ def aiaprep(aiamap):
 
     # Target scale is 0.6 arcsec/pixel, but this needs to be adjusted if the map
     # has already been rescaled.
-    if (aiamap.scale.x/0.6).round() != 1.0*u.arcsec and aiamap.data.shape != (4096, 4096):
-        scale = (aiamap.scale.x/0.6).round() * 0.6*u.arcsec
+    if (aiamap.scale.lon/0.6).round() != 1.0*u.arcsec and aiamap.data.shape != (4096, 4096):
+        scale = (aiamap.scale.lon/0.6).round() * 0.6*u.arcsec
     else:
         scale = 0.6*u.arcsec # pragma: no cover # can't test this because it needs a full res image
-    scale_factor = aiamap.scale.x / scale
+    scale_factor = aiamap.scale.lon / scale
 
     tempmap = aiamap.rotate(recenter=True, scale=scale_factor.value, missing=aiamap.min())
 
@@ -56,7 +56,8 @@ def aiaprep(aiamap):
     # crpix1 and crpix2 will be equal (recenter=True), as aiaprep does not work with submaps
     center = np.floor(tempmap.meta['crpix1'])
     range_side = (center + np.array([-1, 1]) * aiamap.data.shape[0] / 2) * u.pix
-    newmap = tempmap.submap(range_side, range_side)
+    newmap = tempmap.submap(u.Quantity([range_side[0], range_side[0]]),
+                            u.Quantity([range_side[1], range_side[1]]))
 
     newmap.meta['r_sun'] = newmap.meta['rsun_obs'] / newmap.meta['cdelt1']
     newmap.meta['lvl_num'] = 1.5